ENH: initial version of doxygen config file
This doxygen config file has been created with Doxygen 1.7.1 under
Windows. But it also seems to work with older Doxygen versions on our Linux
machines. (I tried it on mantaray via "doxygen doxygen_config", Doxygen
version 1.5.6.) The only directory which is currently browsed is
"../kernel" and its subfolders (path is given relativ to the place of this
config file). The list of directories has to be extended as soon as source code
appears at other places.