ENH: added a "pseudo application"
Added a "pseudo application" for two reasons:

1) If one wants to test something which is currently not possible with the
test/build system, it useful to have one main program which can be
compiled and started independent of the test/build system.

(Current problem: I don't get the test system running when using MPICH.
Until someone explains what is going on (Am I doing something wrong? Or
is there a flaw in the test/build system?), I can use this pseudo app.)

2) When we start thinking about how "user code" is "hooked" into the program
flow (the application programmer will most likely not have control over
the the main() method, but works in some derived class which provides
necessary entry points into the program flow), we can use this pseudo app
(or a copy of it) as a playground.

For the time being, however, you can simply ignore this file.

Signed-off-by: Hilmar Wobker <hilmar.wobker@math.tu-dortmund.de>