ENH: subgroup ranks are now sent via self-defined MPI datatype
The code was already there, but inactive since it didn't work.
As a workaround, the 2D array of subgroup ranks was copied to
a 1D array, which was then sent to the non-coordinator processes.

Now, the code which defines a MPI datatype representing the 2D array
finally works (base address was wrong). The workaround code is removed.

Lesson learned: Don't trust code examples you find on the internet...

Signed-off-by: Hilmar Wobker <hilmar.wobker@math.tu-dortmund.de>