FIX: Previous commit broke compiler flag defaults
In stupid cmake, variables are untyped. So don't expect string functions to
work reliably on variables that the user might or might not have initialised.
In other words, this does not work if the user does not pass in CXXFLAGS via -D:
  if (CXXFLAGS STREQUAL "") do_something endif ()
Why? Because under the hood, CXXFLAGS is a boolean (or an int or something),
as long as no value has been assigned. In other words, all variables are typed
only after their first assignment. So as a workaround, this works
if (NOT CXXFLAGS STREQUAL "")
  do_nothing
else ()
  do_something
endif ()

Signed-off-by: Dominik Goeddeke <dominik.goeddeke@math.tu-dortmund.de>