ENH: implemented proper compiler version detection
This resolves issue 34. Kudos to Sven for the suggested improvements (more
than a month ago). The idea of this (functionally equivalent) version is
to use cmake builtins instead of string arithmetic to figure out if a given
compiler is new enough to compile our code.
Also moved another resolved issue from the open to the closed list, pointless
but good for morale when the TODO list shortens.

Signed-off-by: Dominik Goeddeke <dominik.goeddeke@math.tu-dortmund.de>