BUG: added locale parameter
The C++03-Standard std::toupper and std::tolower functions take two arguments with the second one being of type std::locale. All our supported compilers except for the MSVC seem to have an additional implementation which works with only one argument and - most likely - using a default locale. In the case of the MSVC compiler now the two-argument version is called, where the standard ASCII-Clocale given by std::locale::classic() is used as the second argument.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>