COMP: removed/changed -std=c++0x compiler parameter
The '-std=c++0x' compiler option has been enabled for the GNU, Intel and Open64 compilers, however, only when compiling in DEBUG mode. Since the C++11 support is currently experimental for all compilers, this option lead to fishy behaviour. To improve consistency between compilers, the following steps have been taken:

- The GNU compiler now has the '-std=c++03' option set in both debug- and opt-mode to enforce the C++03 standard, which is (currently) the previously agreed-upon convention. Please note that this option also disables the GNU C++ extensions to the C++03 standard.
- According to the documentation of the Intel C++ v12.1 compiler, there is no '-std=c++03' or any equivalent option available for the ICC, therefore the '-std=c++0x' option has been removed without replacement.
- For the Open64 compiler, the '-std=c++0x' option has also been removed without replacement.

P.S.: With these configurations the GNU compiler will (most probably) be the most pedantic of our currently used compiler set (GCC,ICC,MSC), therefore it is adviseable to test new code with the GCC first, as it will be the first one to complain.
P.P.S.: On a long shot, we (will) want to migrate to the C++11 standard, as it offers a lot of improvement which is interesting in our context. However, as long as the standard exists only on paper and not in compiler, we will stick to the previous C++03 standard for now.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>