ENH: added Finite-Element subsystems
This commit adds the three subsystems 'trafo', 'space' and 'assembly' to
the FEAST kernel library.

The 'trafo' subsystem defines the interfaces for the transformation from a
reference cell onto the real cells of the mesh. It also implements the
standard first-order trafo for 1D and 2D Hypercube meshes.

The 'space' subsystem defines the interfaces for the actual finite-element
spaces defined on a mesh via a transformation. Currently, the 2D
Lagrange-1 (aka Q1) and the 2D Rannacher-Turek (aka Q1~) element space
implementations for conforming meshes have been implemented.

The 'assembly' subsystem implements class templates for the actual assembly
of operators and functionals into matrices and vectors. Currently, the
assembly of scalar bilinear operators into matrices and scalar linear
functionals into vectors has been implemented. Furthermore, a basic
version of the grid-transfer assembly has been implemented.

Please note that up to now the Finite-Element subsystems do not implement
any sort of boundary condition assembly or communication data structure
setup; these components are to be added in near future.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>