STYLE: replaced defect- and product-tests in lafem
The previous tests used stochastically generated matrices and vectors. This was a nice idea, however, the poor choice of the coefficients led to deviances of several orders of magnitude away from machine exactness, which caused the tests to fail when compiling in optimised mode. In addition to that, the TEST_CHECK_EQUALS macro used the container's operator== to perform the comparison, which doomed the test to fail. The matrices have been replaced by 3-point stars and the multiplicant vectors by sine coefficients, ensuring the resulting vectors have coefficients somewhere around 1.

Although the usage of some sort of stochastic/random generated matrices and vectors is generally preferable to something simple as the 3-point start, one has to ensure that the resulting coefficients are in a somwhat useful range. Generating coefficients like 10^12, as done by the previous versions of these tests, is pretty nasty.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>