STYLE: updated mesh test
The Foundation-Mesh-Test made aggressive use of the TEST_CHECK_* macros which (for some unknown reason) cause serious compile-time issues. Replacing the TEST_CHECK_*(...) macros by 'if(...) return false' statements led to a significant reduction of compile-time.

It is yet to be found out whether the compile time issues arise from the TEST_CHECK macros themselves or whether the cause is a combined effect of the Foundation classes and the test system.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>