DOC: added DOXY() macro
This commit adds the DOXY(x) macro which is defined as an empty macro for the compiler, but whose content expands in the doxygen parser. This macro can e.g. be used to document unused parameters.
Furthermore, up to (the lastest) version 1.8.2, doxygen has a bug which leads to wrong namespace assignment for class declarations, i.e. a declaration

<code>
namespace FEAST
{
  template<typename T_>
  class FooBar;
} // namespace FEAST
</code>

will place the FooBar class template in the root namespace rather than the FEAST namespace in the documentation. This can now be resolved by

<code>
namespace FEAST
{
  template<typename T_>
  class FooBar DOXY({});
} // namespace FEAST
</code>

which 'converts' the declaration to a definition for doxygen's parser.

Furthermore, this commit fixes some minor documentation errors.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>