ENH: added matrix- and vector-mirror classes templates
Two new class templates, MatrixMirror and VectorMirror, have been added to
the LAFEM subsystem of FEAST, which are responsible for gather- and
scatter-operations on matrices as well as primal and dual vectors.

Furthermore, a Dof-Mirror assembly class template has been added to
compute the gather- and scatter-matrix structures necessary for the
matrix- and vector-mirrors.

Also fixed a bunch of documentation errors on the way.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>