ENH: added build-mode tags to base-header
A new namespace called 'Modes' as well as four tags classes 'Serial',
'Parallel', 'Debug' and 'Release' have been added to the base-header.
Furthermore, the base-header now contains two typedefs named 'BuildMode',
which is defined as either 'Modes::Serial' or 'Modes::Parallel' depending
on whether the 'PARALLEL' preprocessor macro is defined, and 'DebugMode',
which is defined as either 'Modes::Debug' or 'Modes::Release'.

These tag classes and typedefs allow the selection of template
specialisations according to the currently selected build-mode without
the need for further #ifdef's.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>