ENH: added Geometry Factory interfaces
This commit adds the Geometry::Factory class template as well as its
specialisations for the mesh implementations. A factory object can be
passed to the constructor of a mesh, where the factory's methods are
called for the initialisation of the mesh's data members.

Furthermore, the StandardRefinery class template has been revised and now
derives from the newly added Factory interface.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>