ENH: added discrete projectors
The new DiscreteVertexProjector class projects a discrete Finite-Element
function onto the vertices of the underlying mesh by computing the mean of
all cells adjacent to the vertex.

The DiscreteCellProjector class projects a discrete Finite-Element
function onto the cells of the underlying mesh by computing the integral
mean of the function over a cell using a cubature rule.

Both classes are primarily indented for visualisation output.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>