COMP: disabled 'unused parameter' warnings for GNU compiler
The 'unused parameter' warning has already been disabled for GNU compiler
in debug mode, now it is also disabled in optimised mode.

Unused parameters often appear when a function needs to implement a
specific interface. There are 2 reasons why in some cases unused parameters
still have to be named:
1. doxygen needs the parameter name for correct documentation
2. debug-mode ASSERTs validate the parameter

The fix in the previous commit arised from the above scenario 2.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>