ENH: added node functionals
The NodeFunctional class templates for Finite-Element spaces have been
added. Together with the DofAssignment class templates, these form a
describtion of a space's dual basis, which can be used for various tasks,
most notably the interpolation of a function into a FE space.
This commit also adds the Interpolator class, which interpolates an
analytic function in a FE space.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>