ENH: stringify overload for nullptr
If the compiler supports the C++11 nullptr, a new overload of the
stringify function is enabled, as the lack of it resulted in ambiguities.

Please note that if the compiler does not have built-in support for
nullptr, it is defined as 0, which is of type int, which implies that the
stringify function will return "0" instead of "nullptr" for a nullptr
argument in this case.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>