ENH: added basic Hessian support
The trafo and space structures have been adjusted to support hessian
matrices and tensors. Moreover, the evaluator interfaces have been adapted
for this purpose. The standard trafo mapping supports evaluation of
hessian tensors, however, currently none of the spaces provides hessian
support. This functionality will be provided lateron.

Moreover, cleaned up the evaluation policies for trafos and spaces, thus
reducing the number of typedefs in them.

The changes of this commit are "under the hood", i.e. no modifications to
tests and applications are required.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>