ENH: added support for Hessian matrices
This commit is a major revisement of the space and trafo subsystems.
Adding functionality of computing basis function Hessians revealed a few
flaws of the previous design, which made this revisement necessary.

This commit also includes several changes in the interface between space,
trafo and assembly, however, it does not change the interface between
those and other high-level subsystems.

Currently, there are no finite element spaces implemented which support
computation of Hessian matrices (as these would evaluate to zero), so the
functionality has to be tested afterwards.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>