ENH: added support for standard C assert
This commit introduces a new user-defineable macro named
'FEAST_STDC_ASSERT', which enforces that a call of the 'ASSERT' macro
defined in <kernel/util/assertion.hpp> will redirect to a call of the
'assert' macro of the standard C library instead of potentially throwing
an exception of type FEAST::Assertion.

When using Visual Studio, this has the advantage that a firing call of
assert will trigger a breakpoint in the runtime library, which allows the
user to attach the debugger to the dying process and jump directly into
the failing assertion, even if the process was not started in the debugger
in the first place.

The assertion test has been adjusted such that it only compiles and
executes if the FEAST_STDC_ASSERT macro is undefined.

Note 1: The Visual Studio build system automatically defines
FEAST_STDC_ASSERT from now on.

Note 2: As an assertion is a non-recoverable error by definition, no code
shall try to catch an exception of type FEAST::Assertion. If you want to
have a 'recoverable' assertion, use another exception class for this task.
The assertion test is the only exception to this rule.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>