ENH: revised DofMapping and DofAssignment classes
Up to now, the DofMapping interface implemented the Adjactor interface.
Unfortunately, this assumption is hard to fulfill for finite element
spaces slightly more complex than Q1.
Therefore, the requirement of implementing the Adjactor interface in the
dof-mapping has been dropped, and instead a DofMappingRenderer class has
been added to the dof_adjacency header, which can render a DofMapping
object to an Adjacency::Graph object.

Moreover, this commit introduces the concept of DofTraits, which simplify
the definition of DofMapping and DofAssignment implementations for more
complex elements.

The changes of this commit are under-the-hood, so no code except for the
finite element implementations themselves has to be updated.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>