ENH: 2D Stokes prototype application
This is the prototypic Stokes application, which has been written by the
end of last year, solving a simple 2D Poiseuille-Flow problem on the unit
square domain. It has been adapted to the changes applied to the kernel
since then and furthermore a few extensions have been made.

For the sake of fancyness, this application uses a geometric multigrid
solver in combination with a pressure Schur-complement/SOR smoother :)

This application is just a 'proof-of-concept' and is not indented as a
starting point for further application development.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>