ENH: adapted Adjacency classes to C++11 move operations
All classes in the Adjacency namespace have been adapted to using C++11
features, most noteably move-constructors.

For each of the adjacency classes, the copy-constructors and copy-assign
operators have been replaced by corresponding move variants. In
consequence, the copy variants are now implicitly deleted.
Every of those classes now comes with a clone() member function allowing
to explicitly create a 'deep' copy of the object, which is then returned
and passed on by means of move operations.

For example, the previously valid copy-ctor of the Graph class was used by

  Graph graph2(graph1); // copy graph1 to graph2

which is now invalid. Depending on whether one wants to explicitly create
a deep copy, one now has to write

  Graph graph2(graph1.clone()); // copy graph1 to graph2

or whether one wants to move the graph1 object, one has to write

  Graph graph2(std::move(graph1)); // move graph1 to graph2

This commit also includes several other changes which have now been
possible and preferable due to the introduction of move-constructors:

* The creation of a Graph from a Colouring is now performed by the new
  Colouring::create_partition_graph() member function instead of the
  Graph(const Colouring&) constructor.

* The inversion constructor of the Permutation class has been replaced by
  an inverse() member function returning the inverse Permutation.

* The Cuthill-McKee class has been revised so that one now calls its
  static member function compute() to compute the permutation instead of
  creating an object of type CuthillMcKee.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>