COMP: workaround for ICC 14.0 compiler bug in Tiny algebra
The Intel C++ 14.0 compiler has a bug leading to segmentation faults in
the Tiny::dot() function when compiling optimised builds. A simple
modification of the underlying for-loop did its job as a workaround for
now.

Once Intel releases the 14.1 version, we should check whether the
workaround can be removed again.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>