ENH: outsourced symbolic mirror and matrix assembly
Up to now, symbolic assembly of mirrors and matrices resulted in an
Adjacency::Graph object, which was then passed to the LAFEM container
constructors. However, due to humanity's eternal nemesis* habits of
choking on C++11 code, this lead to the uncompilability of the LAFEM
subsystem with enabled CUDA backend...

Symbolic matrix assembly is now performed by the
Assembly::SymbolicMatrixAssembler class template rather than the
Space::DofAdjacency class template. Moreover, the corresponding member
function now accepts an (uninitialised) Matrix object as a parameter
rather than as a return type. This is less funky than simply calling the
container's ctor, but should resolve the dependency on the C++11 Adjacency
subsystem.

Moreover, vector/matrix mirrors are now assembled by the
Assembly::MirrorAssembler class template rather than by Space::DofMirror.
The reasons are the same as for symbolical matrix assembly.

* aka Nvidia's CUDA C++ compiler wrapper

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>