SYS: Standalone python dash script
This script is meant to be run on every client, that wants to
participate in the nightly unit test runs.
The branches to be exectuted are fetched from and the results are commited
to a mysql database running on augias.
On the other hand, a new web-based interface is to be deplyoed on
augias, to select the branches to be tested on night and to display the
results of the nightly unit tests.
Furthermore, a proper single-summary email has to be send around every morning
at six o'clock by augias.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>