ENH: added PowerVector and TupleVector class templates
This commit introduces the PowerVector class template, which
implements a composition of n sub-vectors of the same class, where
n is a compile-time constant.

Moreover, the TupleVector class template implements a composition of
several sub-vector, where each one may be of a different class.

Finally, this commit introduces a batch of vector operation tests, which
work on a 'TupleVector<PowerVector<DenseVector,2>,DenseVector>'
meta-vector class combination, therefore testing the DenseVector,
PowerVector and TupleVector operator implementation in one shot.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>