ENH: added fallbacks for math functions
As some third-party high-precision libraries (namely GMP) do not implement
the basic math functions by themselves, a set of fallback implemenations
for the following functions has been added.

In total, the FEAST::Math namespace now offers the following functions:

* max (by if)
* min (by if)
* clamp (by max-min)
* signum (by nested if)
* abs (by nested if)
* sqrt (by Newton iteration)
* sin (by series definition)
* cos (by series definition)
* tan (as quotient of sin and cos)
* exp (by series definition)
* log (by Newton iteration)
* log10 (by base-exchange of log)
* pow (by exp-log definition)
* atan (by series definition)
* atan2 (by Wikipedia definition using atan)
* pi (by Leibniz series)
* asin (by atan definition)
* acos (by asin definition)
* eps (see below)

In the case of 'float', 'double' and 'long double' data types,
corresponding overloads ensure that the std:: versions of those functions
(where applicable) are called instead of the fallback implementations.

Note that these fallback implementations do not produce NaNs and infty's
for invalid input parameters, so e.g. sqrt<double>(-1.0) will return 0.0
rather than NaN. For some of the other functions, e.g. log and atan, the
behaviour is undefined for any 'extreme' or invalid cases.

In the special case of the 'eps' function, which returns the machine
precision constant for some data type, no generic implementation is
provided. Currently, this function always calls the
std::numeric_limits::epsilon() function. For other non-built-in types, a
corresponding overload has to be added.

If any third-party high-precision library is added in future, one should
check if that library offers implementations of the standard math
functions. If so, the library's implementation should be preferred over
our fallback implementations, so corresponding overloads should be added.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>