ENH: added PointstarFactory class templates
This commit introduces the PointstarFactoryFD and PointstarFactoryFE
classes, which can be used to generate CSR matrices representing the
Finite-Differences or Finite-Element pointstars.

The PointstarFactoryFD class works with arbitrary dimensions and e.g.
setting the dimension to 2 will lead to the famous 5 point stencil.

The PointstarFactoryFE class currently works only in 2D.

In addition to the assembly of the matrix, these two classes provide
additional functions which compute

* The largest eigenvalue of the matrix.
* The smallest eigenvalue of the matrix.
* The eigenvector with respect to the smallest eigenvalue of the matrix.
* The spectral condition number of the matrix.

Moreover, the member function 'vector_q2_bubble()' will generate an
d-dimensional Q2-bubble vector, which can be used as a desired solution
for linear systems solving the pointstar matrices.

These classes are very helpful for benchmarking and testing purposes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>