ENH: added several Meta-Matrix class templates
This commit introduces several meta-matrix class templates, which
implement compositions of other matrix classes.

In summary, we have:

* PowerDiagMatrix
A meta-matrix which represents a n-by-n diagonal matrix of sub-matrices
of the same class, where n is a compile-time constant. This class does
not assert that the diagonal blocks are square or of the same dimensions,
although this is usually the case.

* PowerRowMatrix
A meta-matrix which represents a 1-by-n row matrix of sub-matrices of the
same class, where n is a compile-time constant. Obviously, all sub-matrix
blocks must have the same number of rows, although currently this
constraint is not checked by the meta-matrix class template.

* PowerColMatrix
A meta-matrix which represents a n-by-1 column matrix of sub-matrices.
This is the 'transposed equivalent' of the PowerRowMatrix.

* SaddlePointMatrix
A meta-matrix which represents a 2-by-2 saddle-point matrix of other
sub-matrices, i.e. a matrix

                       / A  B \
                       \ D  0 /

where the '0' block is implicit.
All three sub-matrices may be of different type, which is the usual case.
In the example of a Stokes system, one usually has that
- A is a n-by-n PowerDiagMatrix
- B is a n-by-1 PowerColMatrix
- D is a 1-by-n PowerRowMatrix

Obviously, it must hold that 'A.rows() = B.rows()' and 'A.columns() =
D.columns()', which is also checked by the class template.

All of these meta-matrix class templates offer the following
functionality:

* Standard-CTORs
* Move-CTORs and Move-Assign-Ops
* A clone() function to create a deep copy
* Implementations of both 'apply()' function variants to perform a
  Matrix-Vector multiplication with a compatible vector type.

Please note that there currently is no m-by-n 'PowerDenseMatrix' or
'TupleMatrix' implementation.

A corresponding meta-matrix test has been added which tests the following
wicked matrix class:

  SaddlePointMatrix<
    PowerDiagMatrix< SparseMatrixCSR >,
    PowerColMatrix< SparseMatrixELL >,
    PowerRowMatrix< SparseMatrixCOO >
    >

which covers all available matrix types with the exception of the
DenseMatrix class template.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>