ENH: added SAME_TYPE macro to base-header
The new SAME_TYPE(T1,T2) macro evaluates to 'true' at compile-time if T1 and
T2 are of the same type, otherwise to 'false'. Useful for static_asserts.

The new macro is used in the TupleVector and SaddlePointMatrix
meta-containers to ensure that the user does not combine types which are
already incompatible at compile-time.

Also fixed two Index vs size_t issues; business as usual...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>