ENH: added meta-vector mirrors
This commit introduces the PowerMirror and TupleMirror class templates, which
can be used to mirror (possibly nested) PowerVector and TupleVector meta
containers. Note that the buffer vector, which is used as in- and output for
mirroring is always a DenseVector.

There are no 'meta-matrix' mirror classes right now and it is not clear
whether they will be necessary at all. If so, they will be added later.

Side-Note:
The TupleMirror class template releaved a nasty compiler bug in the
MSVC compiler: The compiler is not able to deduce variadic template
arguments in special cases.

Consider the following variadic template code:
<code>
  template<typename T1_, typename... T2_>
  struct A {};

  template<typename T_>
  void foo(A<T_>) {}

  void bar() { foo(A<int>); }
</code>

The code will fail to compile with the error code 'C2784' stating that the
template argument of 'A<T_,>' could not be deduced from 'A<int,>'.

In consequence, the gather/scatter-routines had to be 'patched' for the
MSVC compiler, leading to a #ifdef block around the prototype.

The patch of the above example code would be to replace 'foo' by
<code>
  template<typename T_...>
  void foo(A<T_...>) {}
</code>

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>