ENH: added meta-vector filters
This commit introduces the PowerFilter and TupleFilter class templates,
which can be used to filter (possibly nested) PowerVector and TupleVector
meta containers.

There are no 'meta-matrix' filter functions right now and it is not clear
wherher they will be necessary at all. If so, they will be added later.

Please note that this commit does not yet implemented 'filter chains',
which will also be required...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>