ENH: Enforce IT_/Index separation
This commit strictly separates the datatype of scalar values, describing the container
layout (Index) and arrays of unsigned integers (of any bit-length, given
by the last template parameter IT_)
corresponding to the floatingpoint arrays.
This will hopefully enable feast to avoid the hassle of dealing with
different index datatypes out of the containers itself.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>