ENH: added support for Umfpack solver
This commit introduces the LAFEM::Umfpack class, which wraps around the
UMFPACK library. Currently, it only supports double-prec CSR matrices,
whose internal Index type corresponds to FEAST::Index, however, this may
be either a 32 bit or 64 bit integer type.

This commit also adds three CMD scripts for Windows which can be used to
compile UMFPACK, AMD and SuiteSparse_config from source. A solution for
CMake-based builds under linux have yet to be engineered.

Please note that this commit does not include the necessary source codes
of the third-party libraries due to yet unclear copyright issues.
Therefore, the user has to download and unpack the source tar-balls into
the 'thirdparty' directory by hand for now. We may choose to implements
some type of automatism for this like in FEAT2 in future...

Under Windows, the user has to be perform two following steps to get
Umfpack up and running:
1. Download the UMFPACK, AMD and SuiteSparse_config sources from their
   corresponding homepages and unpack them into the 'thirdparty' directory.
   The supported versions can be found in the 'readme.txt'

2. Execute the 'make_umfpack_win32.cmd' and/or 'make_umfpack_win64.cmd'
   scripts to compile and link the library. Please note that the latter
   requires the 64-bit Visual C++ compiler to be installed, of course.

After that, the visual studio project files will automatically detect the
existence of the compiled UMFPACK library and include it into the linker.
Therefore, no further steps are required from the users side.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>