SYS: Added build-id token 'quadmath'
Using this token lets FEAST link against the libquadmath, shipped with
gcc and comptabile to icc.
In addition, the symbol FEAST_HAVE_QUADMATH is defined via
feast_config.hpp.
The common datatype name to be used is called '__float128' and available
at least in gcc and icc.
Note, that in icc this directly translates to its internal type _Quad,
which will be mentioned in all __float128 related error messages.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>