ENH: added quadmath support
This commit adds support for the quadmath library, which is 'supported'
by both the GCC and ICC compilers. Corresponding specialisations and
overloads for the type_traits, math functions and string functions have
been added. It should now be possible to use the '__float128' data-type
for various computations.

For the ICC compiler, the '__float128' type resolves to the build-in type
'_Quad', which (at least to my knowledge) is not officially supported.
However, the ICC compiles and runs the quadmath code without problems.

Moreover, the __float128 type is now also used in the tiny_algebra,
linear_algebra and math tests. More test, especially for the LAFEM
containers, will be added later.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>