ENH: 2014 patch subsystems scarc and foundation (milestone 2)
This major mere 23k LOCs patch is a prerelease of the
milestone 2 reinvention of kernel subsystems scarc and foundation. It
comprises a lot of features but for convenience, a complete commit log
will be provided with the final push. The purpose of this patch is to
fasten development of the control layer.

Signed-off-by: Markus Geveler <markus.geveler@math.tu-dortmund.de>