ENH: revised TEST_CHECK_EQUAL_WITHIN_EPS macro
The previous implementation of the TEST_CHECK_EQUAL_WITHIN_EPS macro
was using the expression 'std::abs(a-b) < eps' for testing.

This approach had two serious drawbacks:
1. It worked only for data types for which std::abs is overloaded, which
   effectively boils to the standard build-in types.
2. It worked only for signed types, e.g. testing with a=5u, b=6u and
   eps=2u failed, because a-b = 5u-6u = bogus.

The new implementation avoids usage of std::abs by testing for the
expression '(a <= b + eps) && (b <= a + eps)', which is mathematically
equivalent to the previous test modulo the fact that now the test check
for 'less or equal' rather than 'strictly less than'. Also the stringify
function evaluation has been adapted to avoid using std::abs.

The advantage is that this check works for any data type which offers the
'+' and '<=' operators. Moreover, the '-' operator is required for the
stringify output of the absolute difference |a-b|.

The drawback is that there are now 6 instantiations of the expressions 'a'
and 'b' instead of 3. This may increase compile-time if a and/or b are not
simple literals or variable lookups, but complex expressions including
(possibly nested) inline function calls.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>