ENH: added component_invert operation to LAFEM vectors
All vector containers have a new operation called 'component_invert',
which takes another vector x and (optionally) a scalar alpha as arguments
and sets the output vectors components to 'alpha/x[i]'. If the scalar
alpha is not given, it is set to 1.

Currently, only the Algo::Generic implementation for Mem::Main exists -
corresponding specialisation for other Algo/Mem combinations are yet to be
written.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>