ENH: added scale_rows/scale_cols operation to SparseMatrixCSR
The SparseMatrixCSR class has benn enriched with two new operations, which
take another matrix a and a vector x as arguments and perform the following
operation:

* scale_rows: r_ij <- a_ij * x_i
* scale_cols: r_ij <- a_ij * x_j

In combination with the component_invert operation from the previous
commit, this operation is required for the assembly of type-0
prolongation/restriction matrices.

Currently, only the Algo::Generic implementation for SparseMatrixCSR in
Mem::Main exists. These operations should also be provided for other matrix
storage formats and Algo/Mem combinations later.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>