ENH: revised GridTransfer assembly to output weight vector
Up to now, the GridTransfer assembly class assembled the prolongation
matrix directly, which was pretty convenient. In order to support the
assembly of type-0 prolongation matrices required for ScaRC, the grid
transfer assembler now returns a matrix and a corresponding weight vector.

After the assembly, the caller needs to invert the vector components by
the newly added 'component_invert' function and afterwards scale the
matrix rows by the 'scale_rows' function to obtain the final (local)
prolongation matrix.

For the assembly of a type-0 (global) prolongation matrix, the weight
vector needs to be synchronised by an add-operation prior to its
inversion. Note that the matrix returned by the assembler must not be
synchronised in this case.

As a side effect, the grid transfer assembly will also be able to assemble
matrix storage types other than CSR once the corresponding ScatterAxpy
class specialisations have been implemented.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>