ENH: added sub-matrix layout ctors for PowerXYZMatrix
The Power{Diag,Col,Row}Matrix classes have been enriched by a constructor
which takes a SparseLayout<...> object as a parameter and initialises all
sub-matrices with that one layout. This is useful, as in 99% of all use
cases all sub-matrices of a Power{Diag,Col,Row}Matrix have identical
layout by definition.

P.S.:
It is somehow confusing to me that the 'LayoutType' member of
SparseMatrix{COO,CSR,ELL} uses the naming convention for types (and in
fact claims to be a type according to its name) altough it is in fact an
enumeration value...

I suggest to rename the current 'LayoutType' to 'layout_tag', and add a
corresponding typedef (yay!) a la

  typedef SparseLayout<MemType, IndexType, layout_tag> LayoutType;

to the SparseMatrix{COO,CSR,ELL} containers.

P.P.S.:
Changed the declaration of LayoutType from 'const static' to
'static constexptr', as it could not (necessarily) be used as a template
parameter otherwise...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>