COMP: Moved op<<(__float128) to namespace std
gcc/icc search in the first place for operator<< in the namespace of the
used ostream, which defaults to std.
If no op<< is found, they create a vast amount of ambigue candidates by
using conversion constructors from __float128 to int, long, double etc.
Thus, any op<< in namespace FEAST is ignored, as long as ostream relies
in namespace std.
As a sidenode: If op<< exists in std and FEAST, the compilers claim
ambigue overloads of op<< found in both namespaces...

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>