ENH: revised LAFEM Power-Container implementations
The previous implementation used recursive template self-inheritance in a
"base-last" manner for the realisation of various Power meta-containers.
This turned out to be problem, as some compilers then searched for members
of base-classes rather than in the most derived class.

To avoid further complication of this type, the Power containers have been
rewritten to a "first-rest" implementation in analogy to the Tuple containers.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>