ENH: Allow and-wise test parameter combination
Recap: by default, a single test is executed, if any of the parameters
to the test binary matches any backend, precision, index or algo.

If the token 'and' is passed to the test binary, all other parameters
must strictly match a tests trais.

Examples:
./dense_vector-test generic cuda
runs all tests with backends generic or cuda.

./dense_vector-test float and cuda
runs only tests with backend cuda and precision float

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>