ENH: revised BilinearOperator/LinearFunctional::Evaluator
The 'Evaluator' class templates of both the BilinearOperator and the
LinearFunctional classes have been revised. The Evaluators now have a
"set_point" member function, which is called for each evaluation point
prior to subsequent calls of the actual evaluation operator, which has now
been stripped of the transformation data parameter. In consequence, any
'per-point' data required by the evaluator now has to be pre-computed in
the "set_point" function and be stored in member variables of the
evaluator.

The main reason for this change is that the previous implementation had
poor performance when the assembly required a lot of computational effort
per cubature point, but which is independent of the actual test/trial
function pair. One important application are custom non-linear operators,
which require the evaluation of one or more finite element functions in a
cubature point.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>